refactor(coordinator/api): Phase 7.1/7.2 — split consumer.go into per-domain files
Pure-move split of the largest file (deletion-only on the source; no body/
comment/signature changed, so all dispatch/billing/streaming invariants —
ensureMaxTokensBound-before-cost, rate-limit-before-reserve, single
provider.Mu() snapshot, post-commit header/cleanup ordering,
invalidateAllAPIKeyCache before+after on key mutations — are preserved):
  consumer.go  4303 -> 32 lines

New files: request_estimate, responses_translate, response_normalize,
response_build, reservation, models_handler, account_handlers, keys_handler,
inference_stream, inference_nonstream, inference_dispatch (the dispatch
primitives + the chat/generic/completions/anthropic handlers).

gofmt clean; go build ./coordinator/... ./e2e/...; go vet; and
go test ./coordinator/api/... (throwaway Postgres) all green; consumer.go
diff is deletions-only.
