+++ coordinator/api/account_handlers.go
@@ -1,0 +1,178 @@
+package api
+
+import (
+	"fmt"
+	"net/http"
+	"strings"
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/payments"
+	"github.com/eigeninference/d-inference/coordinator/store"
+
+	"github.com/eigeninference/d-inference/coordinator/api/types"
+)
+
+// handleHealth handles GET /health.
+// Returns the coordinator's status and the number of connected providers.
+// This endpoint does not require authentication.
+func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
+	writeJSON(w, http.StatusOK, types.HealthResponse{
+		Status:    "ok",
+		Providers: s.registry.ProviderCount(),
+	})
+}
+
+// handleVersion returns the latest provider CLI version and download URL.
+// Providers call GET /api/version to check if they need to update.
+// If a release is registered in the store, uses that. Otherwise falls back
+// to the hardcoded LatestProviderVersion.
+func (s *Server) handleVersion(w http.ResponseWriter, r *http.Request) {
+	const cacheKey = "api_version:v1"
+	if cached, ok := s.readCache.Get(cacheKey); ok {
+		writeCachedJSON(w, cached)
+		return
+	}
+
+	var resp types.VersionResponse
+	// Try release table first.
+	if release := s.store.GetLatestRelease("macos-arm64"); release != nil {
+		resp = types.VersionResponse{
+			Version:      release.Version,
+			Platform:     release.Platform,
+			Backend:      release.Backend,
+			DownloadURL:  release.URL,
+			BinaryHash:   release.BinaryHash,
+			BundleHash:   release.BundleHash,
+			MetallibHash: release.MetallibHash,
+			Changelog:    release.Changelog,
+		}
+	} else {
+		// Fallback to hardcoded version + coordinator download.
+		scheme := "https"
+		if r.TLS == nil && !strings.Contains(r.Host, "darkbloom.dev") {
+			scheme = "http"
+		}
+		downloadURL := fmt.Sprintf("%s://%s/dl/eigeninference-bundle-macos-arm64.tar.gz", scheme, r.Host)
+		resp = types.VersionResponse{
+			Version:     LatestProviderVersion,
+			DownloadURL: downloadURL,
+		}
+	}
+	s.writeCachedJSONResult(w, cacheKey, time.Minute, resp, "failed to encode version")
+}
+
+// --- payment handlers ---
+
+// handleBalance handles GET /v1/payments/balance.
+// Returns the consumer's current balance in both micro-USD and USD.
+func (s *Server) handleBalance(w http.ResponseWriter, r *http.Request) {
+	consumerKey := consumerKeyFromContext(r.Context())
+	balance := s.ledger.Balance(consumerKey)
+	withdrawable := s.store.GetWithdrawableBalance(consumerKey)
+
+	writeJSON(w, http.StatusOK, types.BalanceResponse{
+		BalanceMicroUSD:      balance,
+		BalanceUSD:           fmt.Sprintf("%.6f", float64(balance)/1_000_000),
+		WithdrawableMicroUSD: withdrawable,
+		WithdrawableUSD:      fmt.Sprintf("%.6f", float64(withdrawable)/1_000_000),
+	})
+}
+
+// handleUsage handles GET /v1/payments/usage.
+// Returns the consumer's inference usage history with per-request costs.
+// Tries in-memory ledger first (has full detail), falls back to store
+// ledger history (persists across restarts but has less detail).
+func (s *Server) handleUsage(w http.ResponseWriter, r *http.Request) {
+	consumerKey := consumerKeyFromContext(r.Context())
+	entries := s.ledger.Usage(consumerKey)
+
+	// If in-memory usage is empty (coordinator restarted), build from
+	// the persisted usage table which has full request details.
+	if len(entries) == 0 {
+		usageRecords := s.store.UsageByConsumer(consumerKey)
+		for _, u := range usageRecords {
+			jobID := u.RequestID
+			if jobID == "" {
+				jobID = u.ProviderID
+			}
+			entries = append(entries, payments.UsageEntry{
+				JobID:            jobID,
+				Model:            u.Model,
+				PromptTokens:     u.PromptTokens,
+				CompletionTokens: u.CompletionTokens,
+				CostMicroUSD:     u.CostMicroUSD,
+				Timestamp:        u.CreatedAt,
+			})
+		}
+	}
+
+	writeJSON(w, http.StatusOK, types.UsageResponse{
+		Usage: entries,
+	})
+}
+
+// handleProviderEarnings handles GET /v1/provider/earnings?wallet=0x...
+//
+// Returns the provider's balance and payout history.
+// No API key auth required — providers identify by provider address.
+func (s *Server) handleProviderEarnings(w http.ResponseWriter, r *http.Request) {
+	wallet := r.URL.Query().Get("wallet")
+	if wallet == "" {
+		wallet = r.Header.Get("X-Provider-Wallet")
+	}
+	if wallet == "" {
+		writeJSON(w, http.StatusBadRequest, errorResponse("invalid_request_error", "wallet address required (query param ?wallet=0x... or X-Provider-Wallet header)"))
+		return
+	}
+
+	// Look up balance by provider address
+	balance := s.ledger.Balance(wallet)
+	history := s.ledger.LedgerHistory(wallet)
+	payouts := s.ledger.AllPayouts()
+
+	// Filter payouts to this wallet
+	var walletPayouts []payments.Payout
+	var totalEarned int64
+	var totalJobs int
+	for _, p := range payouts {
+		if p.ProviderAddress == wallet {
+			walletPayouts = append(walletPayouts, p)
+			totalEarned += p.AmountMicroUSD
+			totalJobs++
+		}
+	}
+
+	// If no explicit payout records exist (for example, legacy rows created
+	// before provider_payouts was introduced), reconstruct from persisted
+	// ledger entries with payout type and the wallet as account ID.
+	if len(walletPayouts) == 0 {
+		ledgerEntries := s.store.LedgerHistory(wallet)
+		for _, le := range ledgerEntries {
+			if le.Type == store.LedgerPayout && le.Reference != "" {
+				walletPayouts = append(walletPayouts, payments.Payout{
+					ProviderAddress: wallet,
+					AmountMicroUSD:  le.AmountMicroUSD,
+					JobID:           le.Reference,
+					Timestamp:       le.CreatedAt,
+					Settled:         true,
+				})
+				totalEarned += le.AmountMicroUSD
+				totalJobs++
+			}
+		}
+	}
+
+	if walletPayouts == nil {
+		walletPayouts = []payments.Payout{}
+	}
+
+	writeJSON(w, http.StatusOK, types.ProviderEarningsResponse{
+		BalanceMicroUSD:     balance,
+		BalanceUSD:          fmt.Sprintf("%.6f", float64(balance)/1_000_000),
+		TotalEarnedMicroUSD: totalEarned,
+		TotalEarnedUSD:      fmt.Sprintf("%.6f", float64(totalEarned)/1_000_000),
+		TotalJobs:           totalJobs,
+		Payouts:             walletPayouts,
+		Ledger:              history,
+	})
+}
