+++ coordinator/api/request_estimate.go
@@ -1,0 +1,246 @@
+package api
+
+import (
+	"encoding/json"
+	"fmt"
+	"strings"
+)
+
+func intFromRequestValue(v any) (int, bool) {
+	switch x := v.(type) {
+	case int:
+		return x, true
+	case int32:
+		return int(x), true
+	case int64:
+		return int(x), true
+	case float64:
+		return int(x), true
+	case json.Number:
+		n, err := x.Int64()
+		if err != nil {
+			return 0, false
+		}
+		return int(n), true
+	default:
+		return 0, false
+	}
+}
+
+// approximateTokenCount returns a rough token estimate for routing and queue
+// admission. The len/4 heuristic is a reasonable average for English text
+// with GPT-style BPE tokenizers. This value feeds into the scheduler's
+// capacity checks (pendingTokenBudget, freeMemoryAdmits) where a tighter
+// estimate produces better routing decisions.
+//
+// For billing reservation (where underestimation causes provider shortfall),
+// use approximateTokenCountUpperBound instead.
+func approximateTokenCount(v any) int {
+	if v == nil {
+		return 0
+	}
+	switch x := v.(type) {
+	case string:
+		if x == "" {
+			return 0
+		}
+		tokens := len(x) / 4
+		if tokens < 1 {
+			tokens = 1
+		}
+		return tokens
+	default:
+		b, err := json.Marshal(v)
+		if err != nil {
+			return 0
+		}
+		tokens := len(b) / 4
+		if tokens < 1 {
+			tokens = 1
+		}
+		return tokens
+	}
+}
+
+// approximateTokenCountUpperBound returns a guaranteed upper bound on the
+// number of tokens a BPE tokenizer would produce for v. Every BPE vocabulary
+// starts with one token per byte and can only merge, so len(text) >= tokens
+// for any model family, any language, forever. This is used only for billing
+// reservation to ensure the pre-flight debit always covers the actual cost.
+//
+// Using len(text) over-reserves by ~3-4x on average for English prose, but
+// the difference is refunded immediately after inference completes, so
+// consumers are never overcharged — they only need sufficient balance to
+// cover the reservation hold.
+func approximateTokenCountUpperBound(v any) int {
+	if v == nil {
+		return 0
+	}
+	switch x := v.(type) {
+	case string:
+		return len(x)
+	default:
+		b, err := json.Marshal(v)
+		if err != nil {
+			return 0
+		}
+		return len(b)
+	}
+}
+
+func estimatePromptTokens(parsed map[string]any) int {
+	total := 0
+	if v, ok := parsed["messages"]; ok {
+		total += approximateTokenCount(v)
+	}
+	if v, ok := parsed["input"]; ok {
+		total += approximateTokenCount(v)
+	}
+	if v, ok := parsed["prompt"]; ok {
+		total += approximateTokenCount(v)
+	}
+	if total == 0 {
+		total = approximateTokenCount(parsed)
+	}
+	return total
+}
+
+// estimateBillingPromptTokens returns a guaranteed upper bound on prompt
+// tokens for billing reservation. Uses byte-length (not len/4) so the
+// pre-flight reservation always covers actual cost. This value must NOT
+// be used for routing — see estimatePromptTokens for that.
+func estimateBillingPromptTokens(parsed map[string]any) int {
+	total := 0
+	if v, ok := parsed["messages"]; ok {
+		total += approximateTokenCountUpperBound(v)
+	}
+	if v, ok := parsed["input"]; ok {
+		total += approximateTokenCountUpperBound(v)
+	}
+	if v, ok := parsed["prompt"]; ok {
+		total += approximateTokenCountUpperBound(v)
+	}
+	if total == 0 {
+		total = approximateTokenCountUpperBound(parsed)
+	}
+	return total
+}
+
+func estimateRequestedMaxTokens(parsed map[string]any) int {
+	for _, key := range []string{"max_tokens", "max_completion_tokens", "max_output_tokens"} {
+		if n, ok := intFromRequestValue(parsed[key]); ok && n > 0 {
+			if copies, ok := intFromRequestValue(parsed["n"]); ok && copies > 1 {
+				return n * copies
+			}
+			return n
+		}
+	}
+	if copies, ok := intFromRequestValue(parsed["n"]); ok && copies > 1 {
+		return 256 * copies
+	}
+	return 256
+}
+
+func parseProviderSerialAllowlist(parsed map[string]any) ([]string, bool, error) {
+	var rawValues []any
+	provided := false
+	for _, key := range []string{"provider_serial", "provider_serials"} {
+		v, ok := parsed[key]
+		if !ok {
+			continue
+		}
+		provided = true
+		switch x := v.(type) {
+		case string:
+			rawValues = append(rawValues, x)
+		case []any:
+			rawValues = append(rawValues, x...)
+		default:
+			return nil, true, fmt.Errorf("%s must be a string or array of strings", key)
+		}
+	}
+	if !provided {
+		return nil, false, nil
+	}
+
+	seen := make(map[string]struct{}, len(rawValues))
+	ids := make([]string, 0, len(rawValues))
+	for _, raw := range rawValues {
+		id, ok := raw.(string)
+		if !ok {
+			return nil, true, fmt.Errorf("provider_serials must contain only strings")
+		}
+		id = strings.TrimSpace(id)
+		if id == "" {
+			continue
+		}
+		if _, exists := seen[id]; exists {
+			continue
+		}
+		seen[id] = struct{}{}
+		ids = append(ids, id)
+	}
+	if len(ids) == 0 {
+		return nil, true, fmt.Errorf("provider_serials must include at least one provider serial")
+	}
+	return ids, true, nil
+}
+
+func stripProviderRoutingFields(parsed map[string]any) bool {
+	changed := false
+	for _, key := range []string{"provider_serial", "provider_serials"} {
+		if _, ok := parsed[key]; ok {
+			delete(parsed, key)
+			changed = true
+		}
+	}
+	return changed
+}
+
+// defaultMaxOutputTokens is the ceiling injected into requests that don't set
+// max_tokens. It bounds the worst-case cost of a single inference so the
+// pre-flight balance reservation covers the entire generation; without this
+// cap a consumer could stream output exceeding their reservation and the
+// post-inference charge would fail silently (see GitHub issue #33). Consumers
+// who need longer generations must set max_tokens explicitly and carry the
+// balance to cover it.
+const defaultMaxOutputTokens = 8192
+
+// explicitMaxTokens returns the consumer-specified max output tokens from any
+// of the recognized field names, or 0 if none were set.
+func explicitMaxTokens(parsed map[string]any) int {
+	for _, key := range []string{"max_tokens", "max_completion_tokens", "max_output_tokens"} {
+		if n, ok := intFromRequestValue(parsed[key]); ok && n > 0 {
+			return n
+		}
+	}
+	return 0
+}
+
+// ensureMaxTokensBound injects a max-tokens bound into parsed when the
+// consumer didn't specify any max-tokens field, so the outgoing request to
+// the provider is bounded by the amount we reserve upfront. The bound is
+// the model's max_output_length from the registry (or defaultMaxOutputTokens
+// as fallback). The injected field name depends on the API flavor: Responses
+// API uses max_output_tokens, everything else uses max_tokens. Returns true
+// when an injection occurred, so the caller can re-marshal the outgoing body
+// if needed.
+func ensureMaxTokensBound(parsed map[string]any, isResponsesAPI bool, bound int) bool {
+	if explicitMaxTokens(parsed) > 0 {
+		return false
+	}
+	if isResponsesAPI {
+		parsed["max_output_tokens"] = bound
+	} else {
+		parsed["max_tokens"] = bound
+	}
+	return true
+}
+
+func copyJSONMap(in map[string]any) map[string]any {
+	out := make(map[string]any, len(in))
+	for k, v := range in {
+		out[k] = v
+	}
+	return out
+}
