+++ coordinator/api/reservation.go
@@ -1,0 +1,167 @@
+package api
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/payments"
+	"github.com/eigeninference/d-inference/coordinator/registry"
+	"github.com/eigeninference/d-inference/coordinator/store"
+)
+
+// refundProviderExtra refunds the provider-specific surcharge charged on top of
+// the shared base reservation when an attempt is abandoned. It is idempotent:
+// after refunding it resets ReservedMicroUSD to the base so a second call (or a
+// later settlement) cannot double-refund. The shared base is never refunded
+// here — that is handled once by refundReservation (full failure) or by the
+// winning attempt's settlement.
+func (s *Server) refundProviderExtra(pr *registry.PendingRequest) {
+	if pr == nil {
+		return
+	}
+	extra := pr.ReservedMicroUSD - pr.BaseReservedMicroUSD
+	if extra <= 0 {
+		return
+	}
+	_ = s.store.Credit(pr.ConsumerKey, extra, store.LedgerRefund, "reservation_extra_refund:"+pr.RequestID)
+	pr.ReservedMicroUSD = pr.BaseReservedMicroUSD
+	s.ddIncr("billing.reservation_extra_refunds", []string{"model:" + pr.Model})
+}
+
+// reservationCost is the pre-flight worst-case cost for a text inference
+// request. It mirrors the platform-price branch of handleComplete's billing
+// so the reservation covers any platform-level custom price for the model;
+// without this, a platform override above the built-in default would leave
+// the reservation short and the post-inference clamp would silently
+// undercharge. Provider-specific custom prices are not known until dispatch
+// commits to a provider, so a provider that sets a custom price above the
+// platform rate accepts revenue capped at the reservation.
+func (s *Server) reservationCost(model string, promptTokens, maxTokens int) int64 {
+	customIn, customOut, hasCustom := s.store.GetModelPrice("platform", model)
+	return payments.CalculateCostWithOverrides(model, promptTokens, maxTokens, customIn, customOut, hasCustom)
+}
+
+func (s *Server) refundReservedBalance(pr *registry.PendingRequest, reference string) bool {
+	if pr == nil || pr.ReservedMicroUSD <= 0 {
+		return false
+	}
+	if reference == "" {
+		reference = "reservation_refund:" + pr.RequestID
+	}
+	start := time.Now()
+	finalized, err := pr.FinalizeReservation(func() error {
+		return s.store.Credit(pr.ConsumerKey, pr.ReservedMicroUSD, store.LedgerRefund, reference)
+	})
+	if err != nil {
+		s.logger.Error("failed to refund reservation",
+			"request_id", pr.RequestID,
+			"consumer_key", pr.ConsumerKey,
+			"reserved_micro_usd", pr.ReservedMicroUSD,
+			"error", err,
+		)
+		return false
+	}
+	if !finalized {
+		return false
+	}
+	s.ddIncr("billing.reservation_refunds", []string{"model:" + pr.Model})
+	s.ddHistogram("store.credit.latency_ms", float64(time.Since(start).Milliseconds()), []string{"op:reservation_refund"})
+	return true
+}
+
+// estimateRetryAfter returns a suggested wait time in seconds before retrying
+// a request for the given model. Based on queue depth as a rough proxy for
+// fleet backlog. OpenRouter uses the Retry-After header to schedule retries.
+func (s *Server) estimateRetryAfter(model string) int {
+	queueDepth := s.registry.Queue().QueueSize(model)
+	if queueDepth == 0 {
+		return 2 // Light load, retry soon
+	}
+	// Rough estimate: each queued request takes ~3 seconds to drain.
+	estimate := queueDepth * 3
+	if estimate < 2 {
+		estimate = 2
+	}
+	if estimate > 30 {
+		estimate = 30
+	}
+	return estimate
+}
+
+func providerHasPayoutDestination(provider *registry.Provider) bool {
+	if provider == nil {
+		return false
+	}
+	provider.Mu().Lock()
+	defer provider.Mu().Unlock()
+	return provider.AccountID != ""
+}
+
+func providerPricingKeys(provider *registry.Provider) string {
+	if provider == nil {
+		return ""
+	}
+	provider.Mu().Lock()
+	defer provider.Mu().Unlock()
+	return provider.AccountID
+}
+
+func (s *Server) providerReservationCost(provider *registry.Provider, model string, promptTokens, maxTokens int) int64 {
+	accountID := providerPricingKeys(provider)
+	if accountID != "" {
+		customIn, customOut, hasCustom := s.store.GetModelPrice(accountID, model)
+		if hasCustom {
+			return payments.CalculateCostWithOverrides(model, promptTokens, maxTokens, customIn, customOut, true)
+		}
+	}
+	return s.reservationCost(model, promptTokens, maxTokens)
+}
+
+// isServiceConsumer reports whether the account is a service/wholesale account
+// (e.g. OpenRouter). Such accounts are billed at the advertised platform price,
+// so the provider-price reservation top-up and provider custom pricing are
+// skipped for them. A failed lookup falls back to false (normal consumer).
+func (s *Server) isServiceConsumer(accountID string) bool {
+	if accountID == "" {
+		return false
+	}
+	if u, err := s.store.GetUserByAccountID(accountID); err == nil && u != nil {
+		return u.Role == store.RoleService
+	}
+	return false
+}
+
+func (s *Server) reserveAdditionalForProvider(pr *registry.PendingRequest, provider *registry.Provider) (int64, error) {
+	if pr == nil {
+		return 0, fmt.Errorf("pending request is required")
+	}
+	// Service/wholesale consumers are billed at the platform price at
+	// settlement, so don't top the reservation up to a provider's higher custom
+	// price — the base platform reservation already covers the actual charge.
+	if s.isServiceConsumer(pr.ConsumerKey) {
+		return pr.ReservedMicroUSD, nil
+	}
+	required := s.providerReservationCost(provider, pr.Model, pr.EstimatedPromptTokens, pr.RequestedMaxTokens)
+	if required <= pr.ReservedMicroUSD {
+		return pr.ReservedMicroUSD, nil
+	}
+	// Per-key spend cap re-check against the provider-specific total: the
+	// initial cap check only saw the platform reservation, so a provider whose
+	// custom price exceeds it could otherwise push a capped key over its limit
+	// in a single request. Treat a cap breach like insufficient funds so the
+	// caller excludes this provider (a cheaper one may still fit) and, if none
+	// fit, the request fails with 402. Checked BEFORE charging the top-up.
+	if pr.KeyID != "" && pr.KeyLimitMicroUSD != nil {
+		since := store.KeySpendWindowStart(pr.KeyLimitReset, time.Now())
+		if s.store.KeySpendSince(pr.KeyID, since)+required > *pr.KeyLimitMicroUSD {
+			return pr.ReservedMicroUSD, store.ErrInsufficientBalance
+		}
+	}
+	extra := required - pr.ReservedMicroUSD
+	if err := s.ledger.Charge(pr.ConsumerKey, extra, "reserve:"+pr.ConsumerKey); err != nil {
+		return pr.ReservedMicroUSD, err
+	}
+	pr.ReservedMicroUSD = required
+	s.ddHistogram("billing.reserved_micro_usd", float64(required), []string{"model:" + pr.Model})
+	return required, nil
+}
