+++ coordinator/api/response_build.go
@@ -1,0 +1,317 @@
+package api
+
+import (
+	"encoding/json"
+	"fmt"
+	"strings"
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+	"github.com/google/uuid"
+
+	"github.com/eigeninference/d-inference/coordinator/api/types"
+)
+
+// extractedMessage holds the reconstructed assistant message from SSE chunks,
+// including text content, reasoning, and any tool calls.
+type extractedMessage struct {
+	Content   string           `json:"content"`
+	Reasoning string           `json:"reasoning,omitempty"`
+	ToolCalls []map[string]any `json:"tool_calls,omitempty"`
+}
+
+// extractMessage parses SSE data lines and reconstructs the full assistant
+// message from streaming chunks, including content, reasoning, and tool_calls.
+func extractMessage(chunks []string) extractedMessage {
+	var contentBuilder strings.Builder
+	var reasoningBuilder strings.Builder
+	// Tool calls are indexed — accumulate argument fragments by index.
+	toolCallMap := map[int]map[string]any{}
+
+	for _, chunk := range chunks {
+		line := strings.TrimPrefix(chunk, "data: ")
+		line = strings.TrimSpace(line)
+		if line == "" || line == "[DONE]" {
+			continue
+		}
+
+		var parsed map[string]json.RawMessage
+		if err := json.Unmarshal([]byte(line), &parsed); err != nil {
+			continue
+		}
+
+		choicesRaw, ok := parsed["choices"]
+		if !ok {
+			continue
+		}
+		var choices []struct {
+			Delta struct {
+				Content          string `json:"content"`
+				Reasoning        string `json:"reasoning"`
+				ReasoningContent string `json:"reasoning_content"`
+				ToolCalls        []struct {
+					Index    int    `json:"index"`
+					ID       string `json:"id,omitempty"`
+					Type     string `json:"type,omitempty"`
+					Function struct {
+						Name      string `json:"name,omitempty"`
+						Arguments string `json:"arguments,omitempty"`
+					} `json:"function,omitempty"`
+				} `json:"tool_calls,omitempty"`
+			} `json:"delta"`
+			Message struct {
+				Content          string `json:"content"`
+				Reasoning        string `json:"reasoning"`
+				ReasoningContent string `json:"reasoning_content"`
+			} `json:"message"`
+			FinishReason *string `json:"finish_reason"`
+		}
+		if err := json.Unmarshal(choicesRaw, &choices); err != nil {
+			continue
+		}
+
+		for _, c := range choices {
+			if c.Delta.Content != "" {
+				contentBuilder.WriteString(c.Delta.Content)
+			} else if c.Message.Content != "" {
+				contentBuilder.WriteString(c.Message.Content)
+			}
+			if c.Delta.Reasoning != "" {
+				reasoningBuilder.WriteString(c.Delta.Reasoning)
+			} else if c.Delta.ReasoningContent != "" {
+				reasoningBuilder.WriteString(c.Delta.ReasoningContent)
+			} else if c.Message.Reasoning != "" {
+				reasoningBuilder.WriteString(c.Message.Reasoning)
+			} else if c.Message.ReasoningContent != "" {
+				reasoningBuilder.WriteString(c.Message.ReasoningContent)
+			}
+			for _, tc := range c.Delta.ToolCalls {
+				existing, ok := toolCallMap[tc.Index]
+				if !ok {
+					existing = map[string]any{
+						"index": tc.Index,
+						"function": map[string]any{
+							"arguments": "",
+						},
+					}
+					toolCallMap[tc.Index] = existing
+				}
+				if tc.ID != "" {
+					existing["id"] = tc.ID
+				}
+				if tc.Type != "" {
+					existing["type"] = tc.Type
+				}
+				fn := existing["function"].(map[string]any)
+				if tc.Function.Name != "" {
+					fn["name"] = tc.Function.Name
+				}
+				fn["arguments"] = fn["arguments"].(string) + tc.Function.Arguments
+			}
+		}
+	}
+
+	content := contentBuilder.String()
+	reasoning := reasoningBuilder.String()
+	if cleaned, extractedReasoning := stripThinkBlocks(content); extractedReasoning != "" {
+		content = cleaned
+		if strings.TrimSpace(reasoning) != "" {
+			reasoning += "\n\n" + extractedReasoning
+		} else {
+			reasoning = extractedReasoning
+		}
+	}
+	msg := extractedMessage{Content: content, Reasoning: reasoning}
+	if len(toolCallMap) > 0 {
+		msg.ToolCalls = make([]map[string]any, 0, len(toolCallMap))
+		for i := range len(toolCallMap) {
+			if tc, ok := toolCallMap[i]; ok {
+				delete(tc, "index")
+				msg.ToolCalls = append(msg.ToolCalls, tc)
+			}
+		}
+	}
+	return msg
+}
+
+func buildResponsesUsage(promptTokens, completionTokens uint64, reasoningTokens uint64) types.ResponsesUsage {
+	return types.ResponsesUsage{
+		InputTokens:        int(promptTokens),
+		InputTokensDetail:  types.ResponsesUsageDetail{},
+		OutputTokens:       int(completionTokens),
+		OutputTokensDetail: types.ResponsesUsageDetail{ReasoningTokens: int(reasoningTokens)},
+	}
+}
+
+func buildResponsesIncompleteDetails(finishReason string) *types.ResponsesIncompleteDetail {
+	switch finishReason {
+	case "length":
+		return &types.ResponsesIncompleteDetail{Reason: "max_output_tokens"}
+	case "content_filter":
+		return &types.ResponsesIncompleteDetail{Reason: "content_filter"}
+	default:
+		return nil
+	}
+}
+
+func responseItemID(prefix, requestID string, index int) string {
+	return fmt.Sprintf("%s_%s_%d", prefix, strings.ReplaceAll(requestID, "-", ""), index)
+}
+
+func appendResponsesOutputItems(output []any, requestID string, msg extractedMessage) []any {
+	index := len(output)
+	if msg.Reasoning != "" {
+		output = append(output, map[string]any{
+			"type": "reasoning",
+			"id":   responseItemID("rs", requestID, index),
+			"summary": []map[string]any{{
+				"type": "summary_text",
+				"text": msg.Reasoning,
+			}},
+		})
+		index++
+	}
+	if msg.Content != "" || len(msg.ToolCalls) == 0 {
+		output = append(output, map[string]any{
+			"type": "message",
+			"role": "assistant",
+			"id":   responseItemID("msg", requestID, index),
+			"content": []map[string]any{{
+				"type":        "output_text",
+				"text":        msg.Content,
+				"annotations": []any{},
+			}},
+		})
+		index++
+	}
+	for _, tc := range msg.ToolCalls {
+		fn, _ := tc["function"].(map[string]any)
+		callID, _ := tc["id"].(string)
+		if callID == "" {
+			callID = responseItemID("call", requestID, index)
+		}
+		name, _ := fn["name"].(string)
+		args, _ := fn["arguments"].(string)
+		output = append(output, map[string]any{
+			"type":      "function_call",
+			"id":        responseItemID("fc", requestID, index),
+			"call_id":   callID,
+			"name":      name,
+			"arguments": args,
+		})
+		index++
+	}
+	return output
+}
+
+func buildResponsesResponse(requestID, model string, msg extractedMessage, usage protocol.UsageInfo, seSignature, responseHash string) types.ResponsesResponse {
+	reasoningTokens := uint64(0)
+	if msg.Reasoning != "" {
+		reasoningTokens = uint64(usage.CompletionTokens)
+	}
+	resp := types.ResponsesResponse{
+		ID:        "resp_" + strings.ReplaceAll(requestID, "-", ""),
+		Object:    "response",
+		CreatedAt: time.Now().Unix(),
+		Model:     model,
+		Output:    appendResponsesOutputItems(nil, requestID, msg),
+		Usage:     buildResponsesUsage(uint64(usage.PromptTokens), uint64(usage.CompletionTokens), reasoningTokens),
+	}
+	if seSignature != "" {
+		resp.SESignature = seSignature
+		resp.ResponseHash = responseHash
+	}
+	return resp
+}
+
+func firstChoice(resp types.ChatCompletionResponse) *types.ChatCompletionChoice {
+	if len(resp.Choices) == 0 {
+		return nil
+	}
+	return &resp.Choices[0]
+}
+
+func chatUsageToResponsesUsage(resp types.ChatCompletionResponse, reasoning string) types.ResponsesUsage {
+	reasoningTokens := 0
+	if reasoning != "" {
+		reasoningTokens = resp.Usage.CompletionTokens
+	}
+	return buildResponsesUsage(uint64(resp.Usage.PromptTokens), uint64(resp.Usage.CompletionTokens), uint64(reasoningTokens))
+}
+
+func chatCompletionToResponses(resp types.ChatCompletionResponse, requestedModel, seSignature, responseHash string) types.ResponsesResponse {
+	requestID := strings.TrimPrefix(resp.ID, "chatcmpl-")
+	if requestID == "" {
+		requestID = uuid.NewString()
+	}
+	created := int(resp.Created)
+	if created <= 0 {
+		created = int(time.Now().Unix())
+	}
+
+	msg := extractedMessage{}
+	finishReason := ""
+	if choice := firstChoice(resp); choice != nil {
+		finishReason = choice.FinishReason
+		msg.Content = choice.Message.Content
+		msg.Reasoning = choice.Message.Reasoning
+		msg.ToolCalls = choice.Message.ToolCalls
+	}
+
+	r := types.ResponsesResponse{
+		ID:        "resp_" + strings.ReplaceAll(requestID, "-", ""),
+		Object:    "response",
+		CreatedAt: int64(created),
+		Model:     requestedModel,
+		Output:    appendResponsesOutputItems(nil, requestID, msg),
+		Usage:     chatUsageToResponsesUsage(resp, msg.Reasoning),
+	}
+	if finishReason != "" && finishReason != "stop" {
+		r.IncompleteDetail = buildResponsesIncompleteDetails(finishReason)
+	}
+	if seSignature != "" {
+		r.SESignature = seSignature
+		r.ResponseHash = responseHash
+	}
+	return r
+}
+
+func buildNonStreamingResponse(requestID, model string, msg extractedMessage, usage protocol.UsageInfo, seSignature, responseHash string) types.ChatCompletionResponse {
+	message := types.ChatCompletionMessage{
+		Role:    "assistant",
+		Content: msg.Content,
+	}
+	if msg.Reasoning != "" {
+		message.Reasoning = msg.Reasoning
+	}
+
+	finishReason := "stop"
+	if len(msg.ToolCalls) > 0 {
+		message.ToolCalls = msg.ToolCalls
+		finishReason = "tool_calls"
+	}
+
+	resp := types.ChatCompletionResponse{
+		ID:      "chatcmpl-" + requestID,
+		Object:  "chat.completion",
+		Created: time.Now().Unix(),
+		Model:   model,
+		Choices: []types.ChatCompletionChoice{{
+			Index:        0,
+			Message:      message,
+			FinishReason: finishReason,
+		}},
+		Usage: types.ChatCompletionUsage{
+			PromptTokens:     usage.PromptTokens,
+			CompletionTokens: usage.CompletionTokens,
+			TotalTokens:      usage.PromptTokens + usage.CompletionTokens,
+		},
+	}
+
+	if seSignature != "" {
+		resp.SESignature = seSignature
+		resp.ResponseHash = responseHash
+	}
+
+	return resp
+}
