+++ coordinator/api/response_normalize.go
@@ -1,0 +1,197 @@
+package api
+
+import (
+	"encoding/json"
+	"regexp"
+	"strings"
+)
+
+var thinkBlockPattern = regexp.MustCompile(`(?is)<think>(.*?)</think>\s*`)
+
+func normalizeCompleteChatResponse(obj map[string]any, requestedModel string) {
+	if requestedModel != "" {
+		obj["model"] = requestedModel
+	}
+	for _, key := range []string{"system_fingerprint"} {
+		if v, ok := obj[key]; ok && v == nil {
+			delete(obj, key)
+		}
+	}
+	choices, ok := obj["choices"].([]any)
+	if !ok {
+		return
+	}
+	for _, rawChoice := range choices {
+		choice, ok := rawChoice.(map[string]any)
+		if !ok {
+			continue
+		}
+		if message, ok := choice["message"].(map[string]any); ok {
+			normalizeCompleteMessage(message)
+		}
+		if delta, ok := choice["delta"].(map[string]any); ok {
+			normalizeCompleteMessage(delta)
+		}
+	}
+}
+
+func normalizeCompleteMessage(message map[string]any) {
+	var extractedReasoning string
+	if content, ok := message["content"]; !ok || content == nil {
+		message["content"] = ""
+	} else if contentText, ok := content.(string); ok {
+		cleaned, reasoning := stripThinkBlocks(contentText)
+		message["content"] = cleaned
+		extractedReasoning = reasoning
+	}
+
+	if rc, ok := message["reasoning_content"]; ok {
+		if rcText, ok := rc.(string); ok && rcText != "" {
+			mergeReasoningField(message, rcText)
+		}
+		delete(message, "reasoning_content")
+	}
+	if reasoning, ok := message["reasoning"]; ok && reasoning == nil {
+		delete(message, "reasoning")
+	}
+	if extractedReasoning != "" {
+		mergeReasoningField(message, extractedReasoning)
+	}
+	for _, key := range []string{"tool_calls", "refusal"} {
+		if v, ok := message[key]; ok && v == nil {
+			delete(message, key)
+		}
+	}
+}
+
+func mergeReasoningField(message map[string]any, reasoning string) {
+	reasoning = strings.TrimSpace(reasoning)
+	if reasoning == "" {
+		return
+	}
+	if existing, ok := message["reasoning"].(string); ok && strings.TrimSpace(existing) != "" {
+		if existing != reasoning && !strings.Contains(existing, reasoning) {
+			message["reasoning"] = existing + "\n\n" + reasoning
+		}
+		return
+	}
+	message["reasoning"] = reasoning
+}
+
+func stripThinkBlocks(text string) (string, string) {
+	matches := thinkBlockPattern.FindAllStringSubmatch(text, -1)
+	reasoningParts := make([]string, 0, len(matches)+1)
+	found := len(matches) > 0
+	for _, match := range matches {
+		if len(match) > 1 {
+			if part := strings.TrimSpace(match[1]); part != "" {
+				reasoningParts = append(reasoningParts, part)
+			}
+		}
+	}
+	cleaned := thinkBlockPattern.ReplaceAllString(text, "")
+	lower := strings.ToLower(cleaned)
+	if idx := strings.Index(lower, "<think>"); idx >= 0 {
+		found = true
+		if part := strings.TrimSpace(cleaned[idx+len("<think>"):]); part != "" {
+			reasoningParts = append(reasoningParts, part)
+		}
+		cleaned = cleaned[:idx]
+	}
+	if !found {
+		return text, ""
+	}
+	return strings.TrimSpace(cleaned), strings.Join(reasoningParts, "\n\n")
+}
+
+// normalizeSSEChunk fixes fields in SSE chunks to match the OpenAI spec.
+// Some backends (e.g. vllm-mlx) emit "content":null instead of "content":"",
+// and include "usage":null which strict parsers (ForgeCode, Codex) reject
+// because they expect usage to be either absent or a full object.
+func normalizeSSEChunk(chunk string) string {
+	line := strings.TrimPrefix(chunk, "data: ")
+	// Only trigger the expensive JSON parse for fields we actually fix.
+	// "finish_reason":null appears on every chunk but we don't touch it,
+	// so checking for generic ":null" causes unnecessary JSON round-trips.
+	needsNullFix := strings.Contains(line, `"content":null`) ||
+		strings.Contains(line, `"tool_calls":null`) ||
+		strings.Contains(line, `"usage":null`) ||
+		strings.Contains(line, `"reasoning":null`) ||
+		strings.Contains(line, `"reasoning_content":null`) ||
+		strings.Contains(line, `"refusal":null`) ||
+		strings.Contains(line, `"system_fingerprint":null`)
+	needsReasoningDedup := strings.Contains(line, `"reasoning_content"`)
+	if !needsNullFix && !needsReasoningDedup {
+		return chunk
+	}
+
+	var raw map[string]json.RawMessage
+	if err := json.Unmarshal([]byte(line), &raw); err != nil {
+		return chunk
+	}
+
+	changed := false
+
+	// Remove top-level null fields (usage, system_fingerprint, etc.)
+	// ForgeCode expects usage to be absent or a full object, not null.
+	for _, key := range []string{"usage", "system_fingerprint"} {
+		if v, ok := raw[key]; ok && string(v) == "null" {
+			delete(raw, key)
+			changed = true
+		}
+	}
+
+	// Fix null fields inside choices[].delta
+	if choicesRaw, ok := raw["choices"]; ok {
+		var choices []map[string]json.RawMessage
+		if err := json.Unmarshal(choicesRaw, &choices); err == nil {
+			for i, choice := range choices {
+				if deltaRaw, ok := choice["delta"]; ok {
+					var delta map[string]json.RawMessage
+					if err := json.Unmarshal(deltaRaw, &delta); err == nil {
+						for _, field := range []string{"content", "reasoning_content", "reasoning", "refusal"} {
+							if v, ok := delta[field]; ok && string(v) == "null" {
+								delta[field] = json.RawMessage(`""`)
+								changed = true
+							}
+						}
+						if v, ok := delta["tool_calls"]; ok && string(v) == "null" {
+							delta["tool_calls"] = json.RawMessage(`[]`)
+							changed = true
+						}
+						// Emit BOTH "reasoning" and "reasoning_content" so both
+						// AI SDK (reads reasoning_content) and ForgeCode/other
+						// clients (reads reasoning) see reasoning tokens.
+						if _, hasR := delta["reasoning"]; hasR {
+							if _, hasRC := delta["reasoning_content"]; !hasRC {
+								// Only reasoning exists — copy to reasoning_content for AI SDK.
+								delta["reasoning_content"] = delta["reasoning"]
+								changed = true
+							}
+						} else if rc, hasRC := delta["reasoning_content"]; hasRC {
+							// Only reasoning_content exists — add reasoning alias.
+							delta["reasoning"] = rc
+							changed = true
+						}
+						if changed {
+							choices[i]["delta"], _ = json.Marshal(delta)
+						}
+					}
+				}
+			}
+			if changed {
+				raw["choices"], _ = json.Marshal(choices)
+			}
+		}
+	}
+
+	if !changed {
+		return chunk
+	}
+
+	out, err := json.Marshal(raw)
+	if err != nil {
+		return chunk
+	}
+	return "data: " + string(out)
+}
